Move out test of graph!
