+++ src/graph.rs
@@ -16,22 +16,10 @@ macro_rules! graph {
         // See https://qiita.com/ubnt_intrepid/items/dcfabd5b0ae4d4e105da (Japanese)
         enum DummyGraphNew {}
         impl DummyGraphNew {
-            cagra_parser::graph_impl!($scalar, $proc);
+            $crate::parser::graph_impl!($scalar, $proc);
         }
         DummyGraphNew::graph_new()
     }};
-}
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn test_graph_macro() {
-        let g = graph!(f64, {
-            let x = 1.0;
-            let y = 3.0;
-            let z = x + y + 2.0 * x * y;
-        });
-    }
 }
 
 /// Node of the calculation graph.
